refactor: extract to `components/`
